some code changes
